move wikimedia search to utils
