+++ jparty/utils.py
@@ -224,3 +224,26 @@ class DynamicButton(QPushButton, AutosizeWidget):
 
     def flags(self):
         return 0
+    
+def search_wikimedia_image(query):
+    url = "https://en.wikipedia.org/w/api.php"
+    header = {
+        "User-Agent": "J-NoChance/0.1 (trevorspreadbury@gmail.com)"
+    }
+    params = {
+        "action": "query",
+        "format": "json",
+        "prop": "pageimages",
+        "titles": query,
+        "pithumbsize": 500,
+    }
+    response = requests.get(url, params=params, headers=header)
+    if response.status_code == 200:
+        data = response.json()
+        pages = data.get("query", {}).get("pages", {})
+        for page_id, page_data in pages.items():
+            if "thumbnail" in page_data:
+                return page_data["thumbnail"]["source"]
+        return "No image found."
+    else:
+        return f"Error: {response.status_code}"
